lstm_django/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
finance-datareader==0.9.94
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.0
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.22.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.14.1
